chatView: Postpone fetching backlog until mapped
Tracker (or more precisely: threading restrictions imposed by the
underlying sqlite database) becomes a bottleneck when running too
many queries simultaneously. Depending on the number of channels,
this can be quite noticeable when fetching backlogs for all channels
on startup. Avoid this by only fetching the initial backlog when
mapping the chat log for the first time.