mainWindow: Manage active room via an :active-room property
Which room is active is currently controlled by the chatroomManager.
While having that functionality on a singleton is convenient as it
can be accessed easily from anywhere, a global active room doesn't
make sense with multiple windows - after all, the whole point for
allowing additional windows is the ability to display multiple rooms
at the same time. Moving that functionality into the window instead
allows for that use case, and makes sense conceptually - it's the
room that is selected in the window's sidebar, whose chat log is shown
in the window's content area, whose users the window's user popover
shows etc.