app: Split out addActionEntries() utility function
We currently store all actions in the application, which works fine
for a single-window application, even though some actions conceptually
belong to the window. However if we want to support multiple windows,
we'll have to split our actions - in preparation for this, move our
g_action_map_add_action_entries() replacement into Utils.