chatroomManager: Add forEachRoom()
Both the room list in the sidebar and the stack in the main content
area rely on the manager's ::room-added and ::room-removed signals
to keep track of rooms. This relies on no room having been added
before the signals were connected though, which won't be the case
for windows opened later. So add a small forEachRoom() method that
allows components to iterate over the list of added rooms so they
can pick them up.