roomList: Handle accounts immediately when already prepared
When the AccountsMonitor's accountManager is already prepared when
initializing the room list, the ::account-manager-prepared signal
won't be emitted again and we don't set up change notifications or
pick up the list of active accounts. Check for this case and call
the ::account-manager-prepared handler immediately to get proper
initialization in either case.