chatView: Only show notifications for the active window
When we allow opening additional windows, the condition for showing
a notification may be true for multiple views representing the same
room, but it doesn't make sense to notify more than once for the
same message. To ensure that, restrict notifications to the app's
:active-window. Assuming that this is the only window that may have
toplevel focus, this should work correctly in all cases (i.e. we
don't suppress notifications that should be shown or show ones that
should not).