chatView: Add HoverFilterTag
We want to show the newly added UserPopover when clicking a nick in
the chat log. Currently those nicks use a common tag for general
styling and a per-nick tag to represent the online status of the
nick, however we will also want to indicate that the element is
activatable by highlighting it on hover. To allow this effect while
keeping the status color in a per-nick tag, add a HoverFilterTag
class that takes the color from an existing tag and applies an
opacity effect when hovered.

https://bugzilla.gnome.org/show_bug.cgi?id=760853