userTracker: Allow tracking users across rooms
The UserTracker can be used to track the status of users
in a particular room, however sometimes it is more
interesting to know whether a user is online at all (for
example to determine if it is possible to start a private
conversation with them). While we don't have a way to get
that information, we can approximate it by tracking users
across all rooms we have joined.

https://bugzilla.gnome.org/show_bug.cgi?id=760853