userTracker: Add notify-user actions
It is not uncommon that users need to chat with a particular person
who is not connected at the time. As we are already tracking the
online status of nicks, we can support this case by showing a
notification when a particular user comes online. For that purpose,
add a getNotifyActionName() method that returns the name of a stateful
action that controls whether this feature is enabled for a particular
basenick.

https://bugzilla.gnome.org/show_bug.cgi?id=760853