userDetails: Add offline support
Currently the details have a fixed associated TpContact, which is
fine in the case of the user list where rows represent a particular
nickname and are destroyed when that nickname leaves the room.
However when we expose the details in the chat log, nicks are grouped
together by basenick and may be either on- or offline, so allow changing
and unsetting that property, and add a nickname property that is used
as fallback when no contact is available.

https://bugzilla.gnome.org/show_bug.cgi?id=760853